Add typescript to docs
