+++ packages/docs/gatsby-config.js
@@ -6,6 +6,7 @@ module.exports = {
 		author: "@gatsbyjs"
 	},
 	plugins: [
+		"gatsby-plugin-typescript",
 		"gatsby-plugin-sass",
 		"gatsby-plugin-react-helmet",
 		{
