+++ packages/docs/src/components/main.scss
@@ -1,0 +1,1 @@
+@import '~@nice-digital/design-system/scss/nice-design-system';
