+++ packages/docs/src/pages/index.tsx
@@ -1,7 +1,9 @@
 import React from "react";
+import { Hero } from "@nice-digital/design-system";
 
 const IndexPage = () => (
 	<div>
+		<Hero title="Design System">This is the hero</Hero>
 		<p>It might work one day...</p>
 	</div>
 );
